dbi.pc.in: remove unnecessary subdir from include path
It created an unnecessary include path, which was invalid:
`I... syslog-ng-pe-project-6-0/install/include/dbi/dbi`
From code, it is possible to use libdbi headers with or without a "dbi/" prefix.
As it happens in the case of PE7 and PE6.
This patch adds an additional "..../install/include" PATH to include path.

History of this patch:
a48a74c221e44aa90d56c62b8f1fdea63feac609 patch is added and reverted on PE7 branch.
But not on PE6 branch, which was weird as including "dbi/dbi.h" should not work as dbi.pc file only gives us
"include/dbi" and "include/dbi/dbi" PATHs.
For syslog-ng's luck other modules set up the "install/include", so the "#include dbi/dbi.h" worked.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>